send info back to new peer
